+++ src/parse.py
@@ -2,7 +2,7 @@
 Parses email for name, date and cost from transaction text from email
 '''
 #Routes to function based on bank
-def parse(emailText, bank):
+def parseEmail(emailText, bank):
 	if bank == 'Capital One':
 		return parseCapitalOne(emailText)
 	elif bank == 'Chase':
@@ -48,27 +48,3 @@ def parseChase(emailText):
 			date = w
 	name = name[:-1]
 	return date, name, cost
-
-def test_capital_one():
-	bank = 'Capital One'
-	emailFile = open('./transactions/capitalone.txt', 'r')
-	emailText = emailFile.read()
-	date, name, cost = parse(emailText, bank)
-	assert date == '04/17/2019', 'Capital One: date should be 04/17/2019'
-	assert name == 'SUPER CHICKEN', 'Capital One: name should be SUPER CHICKEN'
-	assert cost == '10.38', 'Capital One: cost should be 10.38'
-
-def test_chase():
-	bank = 'Chase'
-	emailFile = open('./transactions/chase.txt', 'r')
-	emailText = emailFile.read()
-	date, name, cost = parse(emailText, bank)
-	assert date == '04/18/2019', 'Chase: date should be 04/18/2019'
-	assert name == 'Amazon.com', 'Chase: name should be Amazon.com'
-	assert cost == '59.35', 'Chase: cost should be 59.35'
-
-
-if __name__ == "__main__":
-	test_capital_one()
-	test_chase()
-	print("Everything passed")
