--- transactions/capitalone.txt
@@ -1,8 +1,0 @@
-
-RE: Account ending in 4843
-
-As requested, we’re notifying you that on 04/17/2019, at SUPER CHICKEN, a pending authorization or purchase in the amount of $10.38 was placed or charged on your Capital One® credit card account.
-
-Note: You’ll receive this notification for both purchases and pending authorizations, such as car rentals, hotel reservations and gas purchases, even if an actual transaction hasn’t taken place.
-
-Please visit your account to view your pending and posted transactions.
